Refactored Code
