+++ src/components/Game.tsx
@@ -13,6 +13,13 @@ import { GameState, Question, GameMode } from '../types/game'; // Import Questio
 import { UserData } from '../types/userData';
 import { OracleButton, OracleModal } from './Oracle';
 import { TrafficObstacle, NUM_OBSTACLES } from './TrafficObstacle';
+import { FuelIcon } from './FuelIcon';
+import { LevelMap } from './LevelMap';
+import { Road } from './Road';
+import { PlayerCar } from './PlayerCar';
+import {MovingAnswerOptions} from './MovingAnswerOptions';
+import { MovingLaneDividers } from './MovingLaneDividers';
+import { PauseButton } from './PauseButton';
 
 
 // Add debug logging utility
@@ -29,194 +36,6 @@ const SWIPE_TIMEOUT = 300; // Maximum time in ms for a swipe
 // Update these constants at the top of the file
 export const LANE_POSITIONS = [-5, 0, 5]; // Make sure these match your desired positions
 
-// Define Road component at the top level
-function Road() {
-  return (
-    <mesh rotation={[-Math.PI / 2, 0, 0]} receiveShadow position={[0, 0, 0]}>
-      <planeGeometry args={[15, 200]} /> {/* Increased length from 100 to 200 */}
-      <meshStandardMaterial color="#333333" />
-    </mesh>
-  );
-}
-
-// Define PlayerCar component
-function PlayerCar({ position, targetPosition, handleCoinCollect }: { 
-  position: [number, number, number], 
-  targetPosition: number,
-  handleCoinCollect: (id: number) => void 
-}) {
-
-  const { scene } = useGLTF(`${process.env.PUBLIC_URL}/models/car.glb`)
-
-  const rigidBodyRef = useRef<RapierRigidBody>(null);
-  const currentPos = useRef(position[0]);
-  const lastUpdateTime = useRef(0);
-  const isMoving = useRef(false);
-  
-  const MIN_UPDATE_INTERVAL = 16;
-  const COLLISION_BOX = { width: 2, height: 1, depth: 4 };
-
-  useFrame((state, delta) => {
-    if (!rigidBodyRef.current) return;
-
-    try {
-      const targetDiff = targetPosition - currentPos.current;
-      
-      if (Math.abs(targetDiff) > 0.01) {
-        if (!isMoving.current) {
-          isMoving.current = true;
-        }
-
-        // Use a smaller lerp factor for smoother movement
-        const lerpFactor = 0.08;
-        currentPos.current = lerp(currentPos.current, targetPosition,  lerpFactor);
-
-        const newPosition = new Vector3(
-          currentPos.current,
-          position[1],
-          position[2]
-        );
-
-        rigidBodyRef.current.setTranslation(newPosition, true);
-        
-        // Smoother tilt effect
-        const tiltAmount = Math.min(Math.max(targetDiff * -0.1, -0.15), 0.15);
-        const newRotation = new THREE.Quaternion().setFromEuler(
-          new THREE.Euler(0, 0, tiltAmount)
-        );
-        rigidBodyRef.current.setRotation(newRotation, true);
-      } else if (isMoving.current) {
-        
-        // Snap to exact position
-        currentPos.current = targetPosition;
-        rigidBodyRef.current.setTranslation(
-          new Vector3(targetPosition, position[1], position[2]),
-          true
-        );
-        rigidBodyRef.current.setRotation(new THREE.Quaternion(), true);
-        isMoving.current = false;
-      }
-    } catch (error) {
-      console.error('Car movement error:', error);
-    }
-  });
-
-  return (
-    <RigidBody 
-      ref={rigidBodyRef} 
-      position={position} 
-      type="kinematicPosition"
-      colliders={false}
-      userData={{ type: 'PlayerCar' }}
-    >
-      <CuboidCollider 
-        args={[1, 1, 2]} // Adjust size to match car's visible size
-        sensor
-        onIntersectionEnter={(e) => {
-          const otherBody = e.other.rigidBody;
-          const otherType = e.other.rigidBodyObject?.userData?.type;
-          
-          console.log('[PlayerCar] Collision detected:', {
-            otherType,
-            playerLane: currentPos.current,
-            otherPos: otherBody?.translation(),
-            time: Date.now()
-          });
-          
-          if (otherType === 'Obstacle') {
-            // Handle obstacle collision directly here
-            console.log('[PlayerCar] Hit obstacle!');
-          }
-        }}
-      />
-
-      {/* Remove debugging visualization in production */}
-      {/* <mesh scale={[2, 2, 4]}>
-        <boxGeometry />
-        <meshBasicMaterial color="blue" wireframe />
-      </mesh> */}
-
-      {/* Rest of the car model rendering */}
-      <group scale={[0.99, 0.99, 0.99]} rotation={[0, Math.PI, 0]}>
-        {scene ? (
-          <primitive object={scene.clone()} />
-        ) : (
-          <mesh castShadow>
-            <boxGeometry args={[COLLISION_BOX.width, COLLISION_BOX.height, COLLISION_BOX.depth]} />
-            <meshStandardMaterial color="red" wireframe />
-          </mesh>
-        )}
-      </group>
-    </RigidBody>
-  );
-}
-
-// Define MovingLaneDividers component
-function MovingLaneDividers({ gameState }: { gameState: GameState }) {
-  const dividerRefs = useRef<Array<{ position: Vector3 }>>([]);
-  const spacing = 10;
-  const numMarkers = 20;
-
-  useFrame((state, delta) => {
-    if (!gameState.isPlaying || gameState.isPaused) return; // Add pause check
-    
-    const moveAmount = GAME_SPEED * gameState.multiplier * delta * 60;
-    
-    dividerRefs.current.forEach(marker => {
-      marker.position.z += moveAmount;
-      if (marker.position.z > 20) {
-        marker.position.z -= spacing * numMarkers;
-      }
-    });
-  });
-
-  return (
-    <>
-      {[-2, 2].map((x, laneIndex) => (
-        Array.from({ length: numMarkers }).map((_, index) => (
-          <mesh
-            key={`${laneIndex}-${index}`}
-            position={[x, 0.01, -spacing * index]}
-            rotation={[-Math.PI / 2, 0, 0]}
-            ref={(ref: any) => {
-              if (ref) {
-                if (!dividerRefs.current[laneIndex * numMarkers + index]) {
-                  dividerRefs.current[laneIndex * numMarkers + index] = ref;
-                }
-              }
-            }}
-          >
-            <planeGeometry args={[0.2, 3]} />
-            <meshStandardMaterial color="white" emissive="white" emissiveIntensity={0.5} />
-          </mesh>
-        ))
-      ))}
-    </>
-  );
-}
-
-// Add PauseButton component near other UI components
-function PauseButton({ isPaused, onClick }: { isPaused: boolean, onClick: () => void }) {
-  return (
-    <button
-      onClick={onClick}
-      className="fixed top-4 right-20 z-50 bg-white/80 hover:bg-white/90 p-2 rounded-full shadow-lg transition-all duration-200 backdrop-blur-sm"
-      aria-label={isPaused ? "Resume game" : "Pause game"}
-    >
-      {isPaused ? (
-        // Play icon
-        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="currentColor">
-          <path d="M8 5v14l11-7z"/>
-        </svg>
-      ) : (
-        // Pause icon
-        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="currentColor">
-          <path d="M6 19h4V5H6v14zm8-14v14h4V5h-4z"/>
-        </svg>
-      )}
-    </button>
-  );
-}
 
 // Inside the Game component, add touch handling
 const initialGameState: GameState = {
@@ -1040,196 +859,5 @@ export default function Game() {
   );
 }
 
-// Update the MovingAnswerOptions component
-function MovingAnswerOptions({ question, onCollision, gameState }: { 
-  question: Question, 
-  onCollision: (isCorrect: boolean) => void,
-  gameState: GameState 
-}) {
-  const optionsGroupRef = useRef<THREE.Group>(null);
-  const hasCollided = useRef(false);
-  const resetPosition = useRef(false);
-  const initialZ = -180;
 
-  useFrame((state, delta) => {
-    if (!gameState.isPlaying || gameState.isPaused) return; // Add pause check
-    
-    if (!optionsGroupRef.current || resetPosition.current) return;
-
-    const currentZ = optionsGroupRef.current.position.z;
-    
-    // Check for collision based on position
-    if (currentZ > -2 && currentZ < 2) {  // Collision zone
-      const playerLane = gameState.currentLane;
-      question.options.forEach((_, index) => {
-        if (index === playerLane && !hasCollided.current) {
-          debugLog('Position-based collision detected', {
-            optionIndex: index,
-            playerLane,
-            optionZ: currentZ
-          });
-          handleCollision(index);
-        }
-      });
-    }
-
-    // Move options forward
-    optionsGroupRef.current.position.z += GAME_SPEED * gameState.speed * gameState.multiplier;
-    
-    if (currentZ > -5 && currentZ < 5) {
-      debugLog('Options near player', {
-        optionsZ: currentZ,
-        playerLane: gameState.currentLane
-      });
-    }
-    
-    // Reset if passed the player
-    if (currentZ > 10) {
-      resetPosition.current = true;
-      optionsGroupRef.current.visible = false;
-      optionsGroupRef.current.position.z = initialZ;
-      
-      setTimeout(() => {
-        if (optionsGroupRef.current) {
-          optionsGroupRef.current.visible = true;
-          resetPosition.current = false;
-          hasCollided.current = false;
-        }
-      }, 500);
-    }
-  });
-
-  const handleCollision = (index: number) => {
-    if (!hasCollided.current && !resetPosition.current) {
-      hasCollided.current = true;
-      const isCorrect = index === question.correctAnswer;
-      
-      debugLog('Processing collision', {
-        isCorrect,
-        index,
-        correctAnswer: question.correctAnswer
-      });
-      
-      if (isCorrect && optionsGroupRef.current) {
-        resetPosition.current = true;
-        optionsGroupRef.current.visible = false;
-      }
-      
-      onCollision(isCorrect);
-    }
-  };
-
-  // Reset refs when question changes
-  useEffect(() => {
-    hasCollided.current = false;
-    resetPosition.current = false;
-    if (optionsGroupRef.current) {
-      optionsGroupRef.current.position.z = initialZ;
-      optionsGroupRef.current.visible = true;
-    }
-  }, [question]);
-
-  return (
-    <group ref={optionsGroupRef} position={[0, 0, initialZ]} name="optionsGroup">
-      {question.options.map((option, index) => {
-        const lanePosition = LANE_POSITIONS[index];
-        return (
-          <RigidBody
-            key={index}
-            position={[lanePosition, 1, 0]}
-            type="fixed"
-            colliders="cuboid"
-            sensor={true}
-          >
-            <CuboidCollider 
-              args={[2, 1.5, 0.5]} // Increased width and height
-              sensor={true}
-            />
-            <mesh>
-              <planeGeometry args={[4, 3]} /> {/* Use planeGeometry for images */}
-              <meshStandardMaterial 
-                map={new THREE.TextureLoader().load(option)} 
-                transparent={true}
-              />
-            </mesh>
-            {/* Remove Text component and use image */}
-          </RigidBody>
-        );
-      })}
-    </group>
-  );
-}
-
-// First, add this SVG component at the top level of the file
-function FuelIcon({ depleted }: { depleted?: boolean }) {
-  return (
-    <svg 
-      width="24" 
-      height="24" 
-      viewBox="0 0 24 24" 
-      fill="currentColor" 
-      xmlns="http://www.w3.org/2000/svg"
-      className={`transition-opacity duration-300 ${
-        depleted 
-          ? 'text-red-400 opacity-30' // Depleted fuel appears faded
-          : 'text-red-500 opacity-100' // Active fuel is fully visible
-      }`}
-    >
-      <path d="M19.77 7.23l.01-.01-3.72-3.72L15 4.56l2.11 2.11c-.94.36-1.61 1.26-1.61 2.33 0 1.38 1.12 2.5 2.5 2.5.36 0 .69-.08 1-.21v7.21c0 .55-.45 1-1 1s-1-.45-1-1V14c0-1.1-.9-2-2-2h-1V5c0-1.1-.9-2-2-2H6c-1.1 0-2 .9-2 2v16h10v-7.5h1.5v5c0 1.38 1.12 2.5 2.5 2.5s2.5-1.12 2.5-2.5V9c0-.69-.28-1.32-.73-1.77zM12 10H6V5h6v5zm6 0c-.55 0-1-.45-1-1s.45-1 1-1 1 .45 1 1-.45 1-1 1z"/>
-    </svg>
-  );
-}
-
-// Add new LevelMap component
-function LevelMap({ onSelectLevel, onBack }: { 
-  onSelectLevel: (level: number) => void,
-  onBack: () => void 
-}) {
-  const levels = [
-    { id: 1, name: "Basics of the Road", difficulty: "Easy" },
-    { id: 2, name: "Traffic Signs", difficulty: "Easy" },
-    { id: 3, name: "Speed Control", difficulty: "Medium" },
-    { id: 4, name: "Complex Intersections", difficulty: "Medium" },
-    { id: 5, name: "Advanced Navigation", difficulty: "Hard" },
-    // Add more levels as needed
-  ];
-
-  return (
-    <div className="absolute inset-0 flex items-center justify-center z-30 bg-black/80">
-      <div className="bg-white p-8 rounded-xl shadow-2xl max-w-4xl w-full mx-4">
-        <div className="flex justify-between items-center mb-6">
-          <h2 className="text-3xl font-bold text-gray-800">Select Level</h2>
-          <button
-            onClick={onBack}
-            className="text-gray-600 hover:text-gray-800 transition-colors"
-          >
-            <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-        
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-          {levels.map((level) => (
-            <button
-              key={level.id}
-              onClick={() => onSelectLevel(level.id)}
-              className="relative group p-6 bg-gray-100 rounded-lg hover:bg-blue-50 transition-all duration-200"
-            >
-              <div className="absolute inset-0 bg-blue-500/0 group-hover:bg-blue-500/5 rounded-lg transition-colors" />
-              <h3 className="text-xl font-semibold mb-2 text-gray-800">{level.name}</h3>
-              <div className={`text-sm px-2 py-1 rounded-full inline-block
-                ${level.difficulty === 'Easy' ? 'bg-green-100 text-green-700' :
-                  level.difficulty === 'Medium' ? 'bg-yellow-100 text-yellow-700' :
-                  'bg-red-100 text-red-700'}`}
-              >
-                {level.difficulty}
-              </div>
-            </button>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-}
